+++ Cassiopee/Post/Post/Stream/StreamRibbon.cpp
@@ -77,7 +77,7 @@ PyObject* K_POST::compStreamRibbon(PyObject* self, PyObject* args)
                     "streamRibbon: vector must be defined by 3 components.");
     return NULL;
   }
-  for (int i = 0; i < PyList_Size(vectorNames); i++)
+  for (Py_ssize_t i = 0; i < PyList_Size(vectorNames); i++)
   {
     PyObject* tpl0 = PyList_GetItem(vectorNames, i);
     if (PyString_Check(tpl0))
